adding custom error handling, modify ProjectController and TaskController

+++ src/main/java/com/itgirl/project/worktracker/contollers/TaskController.java
@@ -4,6 +4,7 @@ import com.itgirl.project.worktracker.models.Task;
 import com.itgirl.project.worktracker.services.TaskService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,32 +17,55 @@ public class TaskController {
     public TaskService taskService;
 
     @GetMapping("/tasks")
-    public List<Task> getTasks() {
-        return taskService.getTasks();
+    public ResponseEntity<List<Task>> getTasks() {
+        List<Task> tasks = taskService.getTasks();
+
+        return new ResponseEntity<>(tasks, HttpStatus.OK);
     }
 
     @GetMapping("/tasks/{id}")
-    public Task getTask(@PathVariable long id) {
-        return taskService.getTask(id);
-    }
-
-    @DeleteMapping("/tasks/{id}")
-    public void removeTask(@PathVariable long id) {
+    public ResponseEntity<Task> getTask(@PathVariable long id) {
         Task task = taskService.getTask(id);
-        taskService.removeTask(task);
+
+        if (task == null)
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
+        return new ResponseEntity<>(task, HttpStatus.OK);
     }
 
     @PostMapping("/tasks")
-    public void addNewTask(@RequestBody Task task) {
+    public ResponseEntity<Task> addTask(@RequestBody Task task) {
         if (!taskService.saveTask(task)) {
             throw new ConflictException();
         }
 
+        return new ResponseEntity<>(task, HttpStatus.CREATED);
     }
 
-    @PutMapping("/tasks")
-    public void updateTask(@RequestBody Task task) {
+    @PutMapping("/tasks/{id}")
+    public ResponseEntity<Task> updateTask(@PathVariable long id, @RequestBody Task task) {
+
+        Task isExists = taskService.getTask(id);
+        if (isExists == null)
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        else if (task == null)
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+
         taskService.saveTask(task);
+
+        return new ResponseEntity<>(task, HttpStatus.OK);
+    }
+
+    @DeleteMapping("/tasks/{id}")
+    public ResponseEntity<Task> removeTask(@PathVariable long id) {
+        Task task = taskService.getTask(id);
+
+        if (task == null)
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
+        taskService.removeTask(task);
+
+        return new ResponseEntity<>(task, HttpStatus.OK);
     }
 
     @ResponseStatus(value = HttpStatus.CONFLICT, reason = "The task with the same id already exists")
